refact: add some new validations on server side

+++ server/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -10,7 +11,13 @@ class ContactController {
   }
 
   async show(request, response) {
-    const contact = await ContactsRepository.findById(request.params.id);
+    const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
+
+    const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
       return response.status(404).json({ error: 'Contact not found.' });
@@ -28,14 +35,23 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required.' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
 
-    if (contactExists) {
-      return response.status(400).json({ error: 'Email is already been taken.' });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if (contactExists) {
+        return response.status(400).json({ error: 'Email is already been taken.' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     return response.status(201).json(contact);
@@ -50,27 +66,37 @@ class ContactController {
       category_id,
     } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
 
-    if (!contactExists) {
-      return response.status(404).json({ error: 'Contact not found!' });
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactEmailExists = await ContactsRepository.findByEmail(email);
+    const contactExists = await ContactsRepository.findById(id);
 
-    if (contactEmailExists && contactEmailExists.id !== id) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    if (!contactExists) {
+      return response.status(404).json({ error: 'Contact not found!' });
+    }
+
+    if (email) {
+      const contactEmailExists = await ContactsRepository.findByEmail(email);
+
+      if (contactEmailExists && contactEmailExists.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     return response.json(contact);
@@ -78,6 +104,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
 
     await ContactsRepository.delete(id);
